requirements.txt
# https://packaging.python.org/en/latest/guides/distributing-packages-using-setuptools/#working-in-development-mode
# instala pacote local em development-mode
# roda $ pip install -e . quando roda # pip install --requirement requirements.txt
# -e, --editable <path/url>
# usar -vvv para máxima verbosidade, quando necessário
# Install a project in editable mode (i.e. setuptools "develop mode") from a local project path or a VCS url.
#
# instalação local não depende da presença de setup.py
-e .

# external requirements
click>=8.1.3
# conflitando no awscli
#Sphinx
Sphinx>=7.0.1
coverage>=7.2.7
# conflitando com Sphinx
#awscli
#awscli>=1.27.151
flake8>=6.0.0
python-dotenv>=0.5.1
# pareceu necessário make_dataset.py
pylance>=0.4.20
flask>=2.3.2
Office365-REST-Python-Client
